move task_info_formatter
